refactor: move NFT selectors to asset-picker NFT tab
